requirements.txt
# Discord API wrapper
dico-api==0.0.22
dico-interaction==0.0.4
dico-command==0.0.5

# Audio node
discodo @ git+https://github.com/kijk2869/discodo.git@master

# Extra for dico
PyNaCl==1.4.0
aiohttp[speedups]

# Extra for bot
rich==10.9.0
humanize==3.11.0
psutil==5.8.0
aioredis==2.0.0
